[Fix] Additional Extractor fixes related to #954 (#956)
* cleanup on fembed class

* additional fixes

* additional cleanup on xstreamcdn extractor
