+++ app/src/main/java/com/lagradost/cloudstream3/extractors/AsianLoad.kt
@@ -24,14 +24,13 @@ class AsianLoad : ExtractorApi() {
                         ), true
                     )
                         .forEach { stream ->
-                            val qualityString = if ((stream.quality ?: 0) == 0) "" else "${stream.quality}p"
                             extractedLinksList.add(
                                 ExtractorLink(
                                     name,
-                                    "$name $qualityString",
+                                    name = name,
                                     stream.streamUrl,
                                     url,
-                                    getQualityFromName(stream.quality.toString()),
+                                    getQualityFromName(stream.quality?.toString()),
                                     true
                                 )
                             )
