+++ app/src/main/java/com/lagradost/cloudstream3/extractors/GenericM3U8.kt
@@ -30,13 +30,12 @@ open class GenericM3U8 : ExtractorApi() {
             ), true
         )
             .map { stream ->
-                val qualityString = if ((stream.quality ?: 0) == 0) "" else "${stream.quality}p"
                 sources.add( ExtractorLink(
                     name,
-                    "$name $qualityString",
+                    name = name,
                     stream.streamUrl,
                     url,
-                    getQualityFromName(stream.quality.toString()),
+                    getQualityFromName(stream.quality?.toString()),
                     true
                 ))
             }
