+++ app/src/main/java/com/lagradost/cloudstream3/extractors/Pelisplus.kt
@@ -56,7 +56,7 @@ class Pelisplus(val mainUrl: String) {
                         callback.invoke(
                             ExtractorLink(
                                 this.name,
-                                if (qual == "null") this.name else "${this.name} - " + qual + "p",
+                                name = this.name,
                                 href,
                                 page.url,
                                 getQualityFromName(qual),
