+++ app/src/main/java/com/lagradost/cloudstream3/extractors/XStreamCdn.kt
@@ -8,6 +8,22 @@ import com.lagradost.cloudstream3.utils.ExtractorApi
 import com.lagradost.cloudstream3.utils.ExtractorLink
 import com.lagradost.cloudstream3.utils.getQualityFromName
 
+class FEmbed: XStreamCdn() {
+    override val name: String = "FEmbed"
+    override val mainUrl: String = "https://www.fembed.com"
+}
+
+class Fplayer: XStreamCdn() {
+    override val name: String = "Fplayer"
+    override val mainUrl: String = "https://fplayer.info"
+}
+
+class FeHD: XStreamCdn() {
+    override val name: String = "FeHD"
+    override val mainUrl: String = "https://fembed-hd.com"
+    override var domainUrl: String = "fembed-hd.com"
+}
+
 open class XStreamCdn : ExtractorApi() {
     override val name: String = "XStreamCdn"
     override val mainUrl: String = "https://embedsito.com"
